added to backend/src/database/controllers/package/upload.ts
L21: Buffer.from(_package.data.Content),

The error TS2322: Type 'Uint8Array<ArrayBufferLike> | null' is not assignable to type 'Buffer<ArrayBufferLike> | null' indicates that the type of newPackage.content is Uint8Array or null, but it is expected to be Buffer or null.

This error likely occurred because the content field in the Prisma schema is defined as Uint8Array or null, but you are trying to assign it to a Buffer type.

To fix this, you need to convert newPackage.content to a Buffer before assigning it to the Content field in the test object.